Tests/LibWeb/Text/input/async-scrolling/blocking-wheel-event-region-hit-testing.html

<!DOCTYPE html>
<script src="../include.js"></script>
<style>
    body {
        margin: 0;
    }

    #transformed {
        position: absolute;
        left: 20px;
        top: 20px;
        width: 100px;
        height: 100px;
        transform-origin: top left;
        transform: translateX(100px);
    }

    #clip {
        position: absolute;
        left: 20px;
        top: 160px;
        width: 100px;
        height: 100px;
        overflow: hidden;
    }

    #clipped-target {
        width: 200px;
        height: 100px;
    }
</style>
<div id="transformed"></div>
<div id="clip">
    <div id="clipped-target"></div>
</div>
<script>
    function printOutputAndDisplayList(output) {
        const displayList = internals.dumpDisplayList();
        for (const line of output)
            println(line);
        println("display list:");
        println(displayList);
    }

    function dump(output, label, x, y) {
        output.push(`${label}: ${internals.asyncScrollingStateBlocksWheelEventAt(x, y)}`);
    }

    test(() => {
        const output = [];
        transformed.addEventListener("wheel", event => {}, { passive: false });
        document.getElementById("clipped-target").addEventListener("wheel", event => {}, { passive: false });

        output.push(`blocking regions: ${internals.asyncScrollingState().blockingWheelEventRegionCount}`);
        dump(output, "transformed visual inside", 130, 30);
        dump(output, "transformed original position", 30, 30);
        dump(output, "clipped visible inside", 30, 170);
        dump(output, "clipped overflow outside clip", 150, 170);
        printOutputAndDisplayList(output);
    });
</script>
