Tests/LibWeb/Text/expected/async-scrolling/blocking-wheel-event-region-hit-testing.txt

blocking regions: 2
transformed visual inside: true
transformed original position: false
clipped visible inside: true
clipped overflow outside clip: false
